added compute_distances() in HiVis

+++ HiVis/HiVis_plot.py
@@ -226,7 +226,7 @@ class PlotVisium:
             if values is None:
                 raise ValueError(f"{what} not found in adata")
             if values.dtype == np.bool_:
-                values = values.astype(int)
+                values = values.astype(float)
             if np.issubdtype(values.dtype, np.number) and not show_zeros:  # Filter values that are 0
                 if np.all(values == 0):
                     raise ValueError(f"{what} is equal to zero in the specified xlim,ylim")
@@ -598,7 +598,7 @@ class PlotAgg:
             if values is None:
                 raise ValueError(f"{what} not found in adata")
             if values.dtype == np.bool_:
-                values = values.astype(int)
+                values = values.astype(float)
             if np.issubdtype(values.dtype, np.number) and not show_zeros:  # Filter values that are 0
                 if np.all(values == 0):
                     raise ValueError(f"{what} is equal to zero in the specified xlim,ylim")
@@ -710,7 +710,7 @@ class PlotAgg:
             if values is None:
                 raise ValueError(f"{what} not found in adata")
             if values.dtype == np.bool_:
-                values = values.astype(int)
+                values = values.astype(float)
             if np.issubdtype(values.dtype, np.number):
                 if not show_zeros:
                     values[values==0] = np.nan
